chanstate: fold channel kv helper files
Fold the channel key and thaw-height helpers into the store files
that own them, instead of keeping separate kv files that do not
represent their own stores.

The open-channel keys and thaw-height helpers now live with
kv_open_channel, closed-channel bucket ownership lives with
kv_close_summary, and commitment/update keys live with kv_commitment.
